demo/requirements.txt
# Demo Dependencies - NO Docker, NO Kafka required
# Just: pip install -r demo/requirements.txt

# Core processing (same as main pipeline)
pyspark==3.5.0
delta-spark==3.1.0

# Live data collection
requests>=2.31.0
sseclient-py>=1.8.0          # Server-Sent Events client (Wikimedia streams)

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0

# Utilities
tabulate>=0.9.0
